feat: split convert.py into 2 modules engine_object and engine_value along with relevant tests
